refact: move lightning and nostr constants to single file each
