+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/settings/ark.tsx
@@ -13,7 +13,7 @@ import SSButton from '@/components/SSButton'
 import SSCheckbox from '@/components/SSCheckbox'
 import SSText from '@/components/SSText'
 import SSTextInput from '@/components/SSTextInput'
-import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/arkServers'
+import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/ark'
 import { useArkServerAccessToken } from '@/hooks/useArkServerAccessToken'
 import SSHStack from '@/layouts/SSHStack'
 import SSMainLayout from '@/layouts/SSMainLayout'
