+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/signer/ark/account/add/index.tsx
@@ -7,7 +7,7 @@ import SSButton from '@/components/SSButton'
 import SSRadioButton from '@/components/SSRadioButton'
 import SSText from '@/components/SSText'
 import SSTextInput from '@/components/SSTextInput'
-import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/arkServers'
+import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/ark'
 import { useArkAccountBuilder } from '@/hooks/useArkAccountBuilder'
 import { useArkBitcoinAccounts } from '@/hooks/useArkBitcoinAccounts'
 import SSMainLayout from '@/layouts/SSMainLayout'
