+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/signer/ark/index.tsx
@@ -8,7 +8,7 @@ import SSActionButton from '@/components/SSActionButton'
 import SSButton from '@/components/SSButton'
 import SSStyledSatText from '@/components/SSStyledSatText'
 import SSText from '@/components/SSText'
-import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/arkServers'
+import { ARK_SERVERS, ARK_SUPPORTED_NETWORKS } from '@/constants/ark'
 import { NOSTR_PRIVACY_MASK } from '@/constants/nostr'
 import { useFetchBitcoinPrice } from '@/hooks/useFetchBitcoinPrice'
 import SSHStack from '@/layouts/SSHStack'
