+++ apps/mobile/constants/ark.ts
