+++ apps/mobile/constants/nostr.ts
@@ -1,4 +1,8 @@
-import type { NostrRelay } from '@/types/models/Nostr'
+import type {
+  Nip46Method,
+  Nip46PermissionPolicy,
+  NostrRelay
+} from '@/types/models/Nostr'
 
 export const DEFAULT_RETRY_CONFIG = {
   baseDelayMs: 1000,
@@ -25,6 +29,41 @@ export const RELAY_PROTOCOL_PREFIX = 'wss://'
 export const NOSTR_FALLBACK_NPUB_COLOR = '#404040'
 /** Short NDK connect when probing whether at least one relay is reachable. */
 export const NOSTR_RELAY_REACHABILITY_TEST_MS = 10000
+export const NIP46_EVENT_KIND = 24133
+
+export const NIP46_SUPPORTED_METHODS: Nip46Method[] = [
+  'connect',
+  'get_public_key',
+  'nip04_decrypt',
+  'nip04_encrypt',
+  'nip44_decrypt',
+  'nip44_encrypt',
+  'ping',
+  'sign_event'
+]
+
+export const NIP46_DEFAULT_PERMISSIONS: Record<
+  Nip46Method,
+  Nip46PermissionPolicy
+> = {
+  connect: 'always_allow',
+  get_public_key: 'always_allow',
+  nip04_decrypt: 'ask',
+  nip04_encrypt: 'ask',
+  nip44_decrypt: 'ask',
+  nip44_encrypt: 'ask',
+  ping: 'always_allow',
+  sign_event: 'ask'
+}
+
+export const NIP46_REQUEST_TIMEOUT_MS = 60000
+
+export const NIP46_NOSTR_CONNECT_PREFIX = 'nostrconnect://'
+
+export const NIP46_SUBSCRIPTION_LOOKBACK_SECONDS = 10
+
+export const NIP46_EVENT_PREVIEW_MAX_LENGTH = 200
+
 export { PRIVACY_MASK as NOSTR_PRIVACY_MASK } from '@/constants/privacy'
 /** Shown in place of hidden secret keys (nsec / seed words) — short form. */
 export const NOSTR_HIDDEN_KEY_MASK = '••••••••••••••••'
