+++ apps/mobile/hooks/useNip46Connect.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { NIP46_DEFAULT_PERMISSIONS } from '@/constants/nip46'
+import { NIP46_DEFAULT_PERMISSIONS } from '@/constants/nostr'
 import { useNip46Store } from '@/store/nip46'
 import type { Nip46Session } from '@/types/models/Nostr'
 import { randomUuid } from '@/utils/crypto'
