+++ apps/mobile/hooks/useNip46SessionManager.ts
@@ -1,5 +1,5 @@
-import { Nip46BunkerService } from '@/api/nip46BunkerService'
-import { NIP46_SUPPORTED_METHODS } from '@/constants/nip46'
+import { Nip46BunkerService } from '@/api/nostr'
+import { NIP46_SUPPORTED_METHODS } from '@/constants/nostr'
 import { useNip46Store } from '@/store/nip46'
 import type {
   Nip46Method,
