+++ apps/mobile/utils/nip46.ts
@@ -1,7 +1,7 @@
 import {
   NIP46_EVENT_PREVIEW_MAX_LENGTH,
   NIP46_NOSTR_CONNECT_PREFIX
-} from '@/constants/nip46'
+} from '@/constants/nostr'
 import { t } from '@/locales'
 import type { Nip46ParsedUri } from '@/types/models/Nostr'
